--- ricardo/ee/fastapi/app/db/init_db.py
@@ -1,4 +1,0 @@
-"""Initial database setup script."""
-
-# Initialise the database.
-# Creates the tables and optionally populates them with initial data.
