+++ ricardo/ee/fastapi_app/app/api/routers/routers.py
