+++ ricardo/ee/fastapi_app/app/core/config.py
@@ -1,0 +1,1 @@
+"""Config file."""
