+++ ricardo/ee/fastapi_app/app/core/security.py
@@ -1,0 +1,1 @@
+"""Security file."""
