+++ ricardo/ee/fastapi_app/app/models/conservation_efforts.py
@@ -3,8 +3,8 @@
 from sqlalchemy import Column, Date, ForeignKey, Integer, String, Unicode
 from sqlalchemy.orm import relationship
 
-from ricardo.ee.fastapi.app.models.base import Base
-from ricardo.ee.fastapi.app.models.parks import Parks
+from ricardo.ee.fastapi_app.app.models.base import Base
+from ricardo.ee.fastapi_app.app.models.parks import Parks
 
 
 class ConservationEfforts(Base):
