+++ ricardo/ee/fastapi_app/app/models/parks_species.py
@@ -3,9 +3,9 @@
 from sqlalchemy import Column, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
-from ricardo.ee.fastapi.app.models.base import Base
-from ricardo.ee.fastapi.app.models.parks import Parks
-from ricardo.ee.fastapi.app.models.species import Species
+from ricardo.ee.fastapi_app.app.models.base import Base
+from ricardo.ee.fastapi_app.app.models.parks import Parks
+from ricardo.ee.fastapi_app.app.models.species import Species
 
 
 class ParksSpecies(Base):
