+++ ricardo/ee/fastapi_app/app/models/parks_visits.py
@@ -3,9 +3,9 @@
 from sqlalchemy import Column, Date, ForeignKey, Integer
 from sqlalchemy.orm import relationship
 
-from ricardo.ee.fastapi.app.models.base import Base
-from ricardo.ee.fastapi.app.models.parks import Parks
-from ricardo.ee.fastapi.app.models.visitors import Visitors
+from ricardo.ee.fastapi_app.app.models.base import Base
+from ricardo.ee.fastapi_app.app.models.parks import Parks
+from ricardo.ee.fastapi_app.app.models.visitors import Visitors
 
 
 class ParkVisits(Base):
