+++ ricardo/ee/fastapi_app/app/schemas/schemas.py
