+++ ricardo/ee/fastapi_app/tests/test_core.py
